move prisma client to it's own project in monorepo
the prisma client is now in its own project in the moon monorepo
in folder `packages/prisma` which is marked as a workspace require-
ment for the server project. The idea is to be able to use the types
in the database schema in the frontend and possibly in other places.