big refactoring
In this commit, we made several changes to the app. Firstly, we removed the React routing feature from the menu. This decision was made because we encountered difficulties in extracting the table ID number and sending it to the backend. By removing the routing functionality, we aim to simplify the codebase and address the issue more effectively in the future.

Additionally, we removed the outlet and navlink components from the app. These components were no longer necessary for the current version and were causing unnecessary complexity.